requirements.txt
# This code runs with python 3.13.5
# Data analysis and visualization
pandas==2.3.1
matplotlib==3.10.5

# Core image processing
pillow==11.3.0
pillow_heif==1.0.0
opencv-python==4.12.0.88
numpy==2.2.6

# OCR
pytesseract==0.3.13

# Dependencies
contourpy==1.3.3
cycler==0.12.1
fonttools==4.59.1
kiwisolver==1.4.9
packaging==25.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
setuptools==78.1.1
six==1.17.0
tzdata==2025.2
wheel==0.45.1
